Get rid of / mapping to add \v and wrapscan
Wrapscan makes it really hard to find the last of something in a file
(or the first) because it just goes right through the end of the file
back to the beginning.  Makes it so i can't just do 100n or whatever.

The search macros are obnoxious because it means I search with one syntax
and I regex with another.  There isn't really any way to make the regex macro
without being too annoying so best to just get rid of both.